+++ packages/ui-prompting/src/types.ts
@@ -120,9 +120,7 @@ export interface CheckboxPromptQuestion<T extends Answers = Answers> extends Che
 }
 
 export type PromptQuestion<T extends Answers = Answers> =
-    | ListPromptQuestion<T>
-    | InputPromptQuestion<T>
-    | CheckboxPromptQuestion<T>;
+    ListPromptQuestion<T> | InputPromptQuestion<T> | CheckboxPromptQuestion<T>;
 
 export interface ValidationResult {
     isValid: boolean;